Add files via upload
Consider the problem statement for an "Online Auction System" to be developed:

New users can register to the system through an online process. By registering a user agrees to abide by different pre-defined terms and conditions as specified by the system. Any registered user can access the different features of the system authorized to him / her, after he authenticates himself through the login screen. An authenticated user can put items in the system for auction. Authenticated users can place bid for an item. Once the auction is over, the item will be sold to the user placing the maximum bid. Payments are to be made by third party payment services, which, of course, is guaranteed to be secure. The user selling the item will be responsible for its shipping. If the seller thinks he's getting a good price, he can, however, sell the item at any point of time to the maximum bidder available.

List down at-least 5 functional and 5 non-functional requirements that could be obtained from the requirements specifications